Move DiskCertManager to pkg/common
